sync: add `Event` primitive
Event is simply based on `OnceCell<()>`. However, to support
`Event: unset`, a special `OnceCell<()>::unset` operation is added.

Unsetting a `Oncecell` is only possible for ZST (and restricted to `()`
for simplicity), but it adds another edge case when awaiting the
initialization: the state can be set to `STATE_UNSET` before awakening
the task. In that case, we use the waker absence (consumed by the
previous awakening) to acknowledge the notification. Therefore, we
guarantee that notification from `OnceCell::set` always make pending
`wait_initialized` complete.